code/Ingestion/services/chunking/v1.0.0/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
tiktoken==0.8.0
langchain-text-splitters==0.3.5
httpx==0.27.0
